+++ TcgPocket/App/src/pages/inventory/modlues/inventory-display-component.tsx
